Fixing time template tests in library
Someone once told me that time was a predator that stalked us all our lives. But I rather believe than time is a companion who goes with us on the journey, and reminds us to cherish every moment because they'll never come again. What we leave behind is not as important as how we lived. After all, Number One, we're only mortal.