Use combined entry in changelog as well as GitHub
Now the same text that goes into the GitHub release notes (and the NuGet
release notes, in the Library version of the build script) will also be
written to CHANGELOG.md and stored in the Git repo. This does mean
prerelease patch notes will appear twice in the CHANGELOG, once in the
prerelease section and once in the actual release. But this duplication
is better than having an empty release commit in the CHANGELOG.