Format release notes with linked version number
When creating release notes for NuGet packages and GitHub releases, we
should add the link reference target to the description text, so that
the link on the NuGet and GitHub pages will be a proper link. (The link
reference target is added to the bottom of the CHANGELOG.md file, so it
wouldn't be in the newly-created entry unless we add it explicitly).