Use a backup file, not Git, to revert changelog
On build failure, we don't use Git reset to revert the changes made by
the UpdateChangelog target. Instead, we use a temporary file (and clean
up after ourselves when we're done). That way we won't lose someone's
changes if they had written release notes into the changelog but not
committed yet.