Attempt to fix Windows build failures
The Windows build runners are failing to delete files in .git/objects
because those files are read-only. (On Linux and OS X, you can delete
read-only files as long as you have write permission to the containing
directory, but Windows handles files differently). This should fix that
issue, and allow us to delete the temporary directories on Windows.